+++ src/main/java/uk/gov/ida/matchingserviceadapter/configuration/ConfigurationConstants.java
@@ -6,4 +6,7 @@ public interface ConfigurationConstants {
 
     String PRODUCTION_IDP_TRUSTSTORE_NAME = "prod_ida_idp_metadata.ts";
     String INTEGRATION_IDP_TRUSTSTORE_NAME = "test_ida_idp_metadata.ts";
+
+    String PRODUCTION_METADATA_TRUSTSTORE_NAME = "prod_metadata_truststore.ts";
+    String INTEGRATION_METADATA_TRUSTSTORE_NAME = "test_metadata_truststore.ts";
 }